refactor: Modal Files Restructure
Refactored modal file systems for ease of reading
