+++ src/components/Modals/JournalModal/JournalModal.js
@@ -1,33 +1,33 @@
 import PropTypes from 'prop-types';
 import { Form } from 'react-final-form';
 import arrayMutators from 'final-form-arrays';
-import { useMutation } from 'react-query';
 import { FormattedMessage } from 'react-intl';
+import { useMutation } from 'react-query';
 
-import { Button, Modal, ModalFooter } from '@folio/stripes/components';
 import { useOkapiKy } from '@folio/stripes/core';
+import { Button, Modal, ModalFooter } from '@folio/stripes/components';
 
-import PartyInfoForm from '../PartyFormSections/PartyInfoForm/PartyInfoForm';
+import JournalInfoForm from '../../JournalFormSections/JournalInfoForm';
 
 const propTypes = {
   showModal: PropTypes.bool,
   setShowModal: PropTypes.func,
-  handlePartyChange: PropTypes.func,
+  handleJournalChange: PropTypes.func,
 };
 
-const PartyModal = ({ showModal, setShowModal, handlePartyChange }) => {
+const JournalModal = ({ showModal, setShowModal, handleJournalChange }) => {
   const ky = useOkapiKy();
 
   const handleClose = () => {
     setShowModal(false);
   };
 
-  const { mutateAsync: postParty } = useMutation(
-    ['ui-oa', 'PartyModal', 'postParty'],
-    (data) => ky.post('oa/party', { json: data }).json().then((res) => {
-        handlePartyChange(res);
-        handleClose();
-      })
+  const { mutateAsync: postJournal } = useMutation(
+    ['ui-oa', 'JournalModal', 'postJournal'],
+    (data) => ky.post('oa/works/citation', { json: data }).json().then((res) => {
+      handleJournalChange(res);
+      handleClose();
+    })
   );
 
   const renderModalFooter = (handleSubmit, formRestart) => {
@@ -35,7 +35,7 @@ const PartyModal = ({ showModal, setShowModal, handlePartyChange }) => {
       <ModalFooter>
         <Button
           buttonStyle="primary"
-          id="party-modal-save-button"
+          id="journal-modal-save-button"
           onClick={() => {
             handleSubmit().then(formRestart);
           }}
@@ -45,7 +45,7 @@ const PartyModal = ({ showModal, setShowModal, handlePartyChange }) => {
         </Button>
         <Button
           buttonStyle="default"
-          id="party-modal-cancel-button"
+          id="journal-modal-cancel-button"
           onClick={() => setShowModal(false)}
         >
           <FormattedMessage id="stripes-components.cancel" />
@@ -56,18 +56,24 @@ const PartyModal = ({ showModal, setShowModal, handlePartyChange }) => {
 
   return (
     <Form
+      // Setting initial values of type to serial instead of select field
+      initialValues={{
+        title: '',
+        type: 'serial',
+        instances: [{ ids: [{ ns: '', id: '' }], subType: '' }],
+      }}
       mutators={arrayMutators}
-      onSubmit={postParty}
+      onSubmit={postJournal}
       render={({ handleSubmit, form: { restart: formRestart } }) => (
         <form onSubmit={handleSubmit}>
           <Modal
             dismissible
             footer={renderModalFooter(handleSubmit, formRestart)}
-            label={<FormattedMessage id="ui-oa.party.newPerson" />}
+            label={<FormattedMessage id="ui-oa.journal.newJournal" />}
             onClose={() => setShowModal(false)}
             open={showModal}
           >
-            <PartyInfoForm />
+            <JournalInfoForm />
           </Modal>
         </form>
       )}
@@ -75,6 +81,6 @@ const PartyModal = ({ showModal, setShowModal, handlePartyChange }) => {
   );
 };
 
-PartyModal.propTypes = propTypes;
+JournalModal.propTypes = propTypes;
 
-export default PartyModal;
+export default JournalModal;
