+++ src/components/Modals/JournalModal/index.js
