+++ src/components/Modals/index.js
@@ -1,0 +1,2 @@
+export { default as PartyModal } from './PartyModal';
+export { default as JournalModal } from './JournalModal';
